Optimizacion y mejoras
-Ligera optimizacion
-Posibilidad de aumentar la historia
-Bucle generacion de enemigos dinamico
-Eliminacion de ColorRGB
-Eliminacion parcial de los avisos (principalmente de casting)

+++ TacticalTakedown/src/Mapa.h
@@ -20,7 +20,7 @@ private: //Para incluir los bordes en lista paredes necesitan ser punteros (Pare
 public:
 	Mapa();
 	//virtual ~Mapa();
-	float getNumero() { return numero; }
+	int getNumero() { return numero; }
 
 	bool agregar(Pared* e);
 	void dibujaBordes();
